+++ blueprints/flexberry-edit-form/snippets/getCellComponent-flexberry-lookup.js
@@ -1,14 +1,12 @@
-﻿        case '<%= master %>':
-          return {
-            componentName: 'flexberry-lookup',
-            componentProperties: {
-              choose: 'showLookupDialog',
-              chooseText: '...',
-              remove: 'removeLookupValue',
-              displayAttributeName: '<%= displayAttributeName %>',
-              required: <%= required %>,
-              relationName: '<%= master %>',
-              projection: '<%= projection %>',
-              autocomplete: true
-            }
+        case '<%= detailModelName %>+<%= master %>':
+          cellComponent.componentProperties = {
+            choose: 'showLookupDialog',
+            chooseText: '...',
+            remove: 'removeLookupValue',
+            displayAttributeName: '<%= displayAttributeName %>',
+            required: <%= required %>,
+            relationName: '<%= master %>',
+            projection: '<%= projection %>',
+            autocomplete: true,
           };
+          break;