+++ tests/dummy/app/controllers/ember-flexberry-dummy-suggestion-edit.js
@@ -57,5 +57,11 @@ export default BaseEditFormController.extend(EditFormControllerOperationsIndicat
     }
 
     return cellComponent;
+  },
+
+  actions: {
+    configurateRow(rowConfig, record) {
+      rowConfig.customClass += 'positive';
+    }
   }
 });